+++ frontend/layouts/default.vue
@@ -116,7 +116,7 @@ export default defineComponent({
           icon: this.$globals.icons.link,
           title: "Import",
           subtitle: "Import a recipe by URL",
-          to: "/recipe/create?tab=url",
+          to: "/recipe/create/url",
           restricted: true,
         },
         { divider: true },
@@ -124,7 +124,7 @@ export default defineComponent({
           icon: this.$globals.icons.edit,
           title: "Create",
           subtitle: "Create a recipe manually",
-          to: "/recipe/create?tab=new",
+          to: "/recipe/create/new",
           restricted: true,
         },
         { divider: true },
