+++ frontend/pages/recipe/create/index.vue
@@ -1,0 +1,18 @@
+<template>
+  <div></div>
+</template>
+
+<script lang="ts">
+import { defineComponent, onMounted, useRouter } from "@nuxtjs/composition-api";
+
+export default defineComponent({
+  setup() {
+    const router = useRouter();
+    onMounted(() => {
+      // Force redirect to first valid page
+      router.replace("/recipe/create/url");
+    });
+    return {};
+  },
+});
+</script>
